Vehicles tab: show total km delivered today instead of order count
- dispatcher/views.py: change orders_today annotation from Count to
  Sum(distance_km) with simplified filter (no delivery join to avoid
  double-counting); keep completed_at + updated_at fallback
- dispatcher/serializers.py: update orders_today field to DecimalField
- AX_Dispatch_Portal.jsx: rename column header to 'Deliveries km' and
  render value with fmtDistance() instead of raw integer

+++ backend/dispatcher/serializers.py
@@ -1037,7 +1037,7 @@ class RelayNodeSerializer(serializers.ModelSerializer):
 
 class VehicleAssetSerializer(serializers.ModelSerializer):
     assigned_rider = serializers.SerializerMethodField()
-    orders_today = serializers.IntegerField(read_only=True, default=0)
+    orders_today = serializers.DecimalField(max_digits=10, decimal_places=2, read_only=True, default=0)
 
     class Meta:
         model = VehicleAsset
